pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.38.33",
    "fastapi>=0.115.12",
    "langgraph>=0.4.8",
    "logfire>=3.18.0",
    "matplotlib>=3.10.3",
    "mcp>=1.9.3",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "pydantic>=2.11.5",
    "pydantic-ai>=0.2.16",
    "pydantic-ai-slim[bedrock]>=0.2.16",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "seaborn>=0.13.2",
    "streamlit>=1.45.1",
    "uvicorn>=0.34.3",
]
